finish refactoring functions
